+++ src/operating_system/tinyos/ec_pal/Timer.h
@@ -8,11 +8,11 @@
 
 void* pal_timer_wire();
 
-//void timer_sleep(uint64_t until);
+//void timer_sleep(uint32_t until);
 typedef struct {
     ec_continuation_t ec_continuation;
 } ec_struct_timer_sleep;
-void ec_pal_timer_sleep(void* handle, uint64_t until);
+void ec_pal_timer_sleep(void* handle, uint32_t until);
 extern ec_struct_timer_sleep ec_state_timer_sleep[];
 ec_struct_timer_sleep* ec_map_timer_sleep();
 
@@ -20,6 +20,6 @@ ec_struct_timer_sleep* ec_map_timer_sleep();
 * Return the current time.
 * @return Current time.
 */
-uint64_t pal_timer_getNow(void* handle);
+uint32_t pal_timer_getNow(void* handle);
 
 #endif
